v4.0

+++ script2.sh
@@ -1,7 +1,7 @@
 #!/bin/bash
 
 if pgrep -a nginx >/dev/null; then
-    echo "nginx"
+    echo "NGINX"
     DIRS=$(sudo find /etc/nginx -type f \( -name '*.crt' -o -name '*.key' -o -name '*.pem' -o -iname '*bundle*' \) 2>/dev/null | xargs -r dirname | sort -u)
 	echo "$DIRS"
     DIR_CONT=$(echo "$DIRS" | wc -l)
@@ -18,12 +18,37 @@ if pgrep -a nginx >/dev/null; then
     else
         echo "Diretório encontrado: $DIRS"
         cd "$DIRS" || exit 1
-        for arq in *.crt *.key *.br.pem; do
+        for arq in *.crt; do
             [ -f "$arq" ] && mv "$arq" "$arq.bak2"
-            echo "backup realizado"
-            novo=$(find /tmp/novo_cert -maxdepth 1 -type f \( -iname '*.crt' -o -iname '*.key' -o -iname '*.br.pem' \) | head -n 1)
+            echo "Backup realizado"
+            novo=$(find /tmp/novo_cert -maxdepth 1 -type f \( -iname '*.crt' \) | head -n 1)
             if [ -n "$novo" ]; then
                 echo "Copiando novo arquivo $novo para $DIRS como  $arq"
+                cp "$novo" "$arq"
+                rm "$novo"
+            else
+                echo "Nenhum novo arquivo disponível para substituir $arq"
+            fi
+        done
+	for arq in *.key; do
+            [ -f "$arq" ] && mv "$arq" "$arq.bak2"
+            echo "Backup realizado"
+            novo=$(find /tmp/novo_cert -maxdepth 1 -type f \( -iname '*.key' \) | head -n 1)
+            if [ -n "$novo" ]; then
+                echo "Copiando novo arquivo $novo para $DIRS como  $arq"
+                cp "$novo" "$arq"
+                rm "$novo"
+            else
+                echo "Nenhum novo arquivo disponível para substituir $arq"
+            fi
+        done
+        for arq in ufpe*.pem; do
+            [ -f "$arq" ] || continue
+            mv "$arq" "$arq.bak2"
+            echo "Backup realizado"
+            novo=$(find /tmp/novo_cert -maxdepth 1 -type f -iname '*.pem' | head -n 1)
+            if [ -n "$novo" ]; then
+                echo "Copiando novo arquivo $novo para $DIRS como $arq"
                 cp "$novo" "$arq"
                 rm "$novo"
             else
@@ -41,8 +66,8 @@ if pgrep -a nginx >/dev/null; then
     fi
 
 elif pgrep -a httpd >/dev/null || pgrep -a apache2 >/dev/null; then
-    echo "apache"
-    DIRS=$(sudo find /etc/httpd /etc/apache2 -type f \( -name '*.crt' -o -name '*.key' -o -name '*.pem' -o -iname '*bundle*' \) 2>/dev/null | xargs -r dirname | sort -u)
+    echo "APACHE/HTTPD"
+    DIRS=$(sudo find /etc/httpd /etc/apache2 -type f \( -name '*.crt' -o -name '*.key' -o -name '*.pem' \) 2>/dev/null | xargs -r dirname | sort -u)
 	echo "$DIRS"
     DIR_CONT=$(echo "$DIRS" | wc -l)
 	echo "$DIR_CONT"
@@ -58,10 +83,10 @@ elif pgrep -a httpd >/dev/null || pgrep -a apache2 >/dev/null; then
     else
         echo "Diretório encontrado: $DIRS"
         cd "$DIRS" || exit 1
-        for arq in *.crt *.key *bundle*; do
+        for arq in *.crt; do
             [ -f "$arq" ] && mv "$arq" "$arq.bak2"
             echo "backup realizado"
-            novo=$(find /tmp/novo_cert -maxdepth 1 -type f \( -iname '*.crt' -o -iname '*.key' \) | head -n 1)
+            novo=$(find /tmp/novo_cert -maxdepth 1 -type f \( -iname '*.crt' \) | head -n 1)
             if [ -n "$novo" ]; then
                 echo "Copiando novo arquivo $novo para $DIRS como  $arq"
                 cp "$novo" "$arq"
@@ -70,6 +95,19 @@ elif pgrep -a httpd >/dev/null || pgrep -a apache2 >/dev/null; then
                 echo "Nenhum novo arquivo disponível para substituir $arq"
             fi
         done
+	for arq in *.key; do
+            [ -f "$arq" ] && mv "$arq" "$arq.bak2"
+            echo "backup realizado"
+            novo=$(find /tmp/novo_cert -maxdepth 1 -type f \( -iname '*.key' \) | head -n 1)
+            if [ -n "$novo" ]; then
+                echo "Copiando novo arquivo $novo para $DIRS como  $arq"
+                cp "$novo" "$arq"
+                rm "$novo"
+            else
+                echo "Nenhum novo arquivo disponível para substituir $arq"
+            fi
+        done
+	
         if sudo httpd -t || apachectl -t; then
             echo "Configuração válida. Reiniciando serviço."
             sudo systemctl restart httpd || sudo systemctl restart apache2 || sudo httpd -k restart || sudo apachectl restart
